replace gang verification with oneliner throwIfNoGang
