+++ src/Gangs/lib/getTaskNames.ts
@@ -1,15 +1,10 @@
 import { BaseServer } from "../../Server/BaseServer";
 import {GangMemberTasks} from "../../Gang";
-import { hasGang } from "./hasGang";
+import {throwIfNoGang} from "./throwIfNoGang";
 
 export function getTaskNames(server: BaseServer, term: any, out:Function, err:Function, args: string[], options:any={type:false, list:false}) {
-    let hasGangAlready:boolean = false;
-    let hasGangOut = (value:boolean)=>{hasGangAlready=value;};
-    hasGang(server, term, hasGangOut, err, [])
-    if(!hasGangAlready){
-        err(`You dont have a gang`);
-        return false;
-    }
+
+    throwIfNoGang(server, term, err);
     for(let task of Object.keys(GangMemberTasks)) out(task);
     return true;
 }
